Renamed general service til utility. Added error handling in house
