+++ src/api/house.js
@@ -1,6 +1,6 @@
 import { addDoc, collection, doc, getDoc, getDocs, updateDoc, deleteDoc  } from "firebase/firestore";
 import { database } from "boot/firebaseConfig";
-import { getReadableTimestamp } from "src/service/general.js";
+import { getReadableTimestamp } from "src/service/utility.js";
 import { ref } from "vue";
 
 const collectionName = "houses";
@@ -8,21 +8,31 @@ const collectionName = "houses";
 //Take a house object and creates a document in the database
 export async function addHouse(house){
   const currentTimestamp = getReadableTimestamp()
-  await addDoc(collection(database, collectionName), {
-    name: house.name,
-    description: house.description,
-    created: currentTimestamp,
-    updated: currentTimestamp
-  })
+  try {
+    await addDoc(collection(database, collectionName), {
+      name: house.name,
+      description: house.description,
+      created: currentTimestamp,
+      updated: currentTimestamp
+    })
+  }catch (e){
+    throw new Error(`Failed to add house with name: ${house.name}\n${e.stack}`)
+  }
 }
 
 //Returns a list of all the houses in the database
 export async function getAllHouses(){
   const houses = ref([]);
-  const querySnapshot = await getDocs(collection(database, collectionName));
-  querySnapshot.forEach((doc) => {
-    houses.value.push(doc.data());
-  });
+  try{
+    const querySnapshot = await getDocs(collection(database, collectionName));
+    querySnapshot.forEach((doc) => {
+      houses.value.push(doc.data());
+    });
+  }
+  catch (e){
+    throw new Error(`Failed to retrieve all houses.\n${e.stack}`)
+  }
+
 
   return houses;
 }
@@ -30,20 +40,33 @@ export async function getAllHouses(){
 //Takes an ID and returns the full house object
 export async function getHouseById(id){
   const docRef = doc(database, collectionName, id);
-  const house = await getDoc(docRef);
-  return house.data();
+  try{
+    const house = await getDoc(docRef);
+    return house.data();
+  }catch (e) {
+    throw new Error(`Failed to retrieve house with id: ${id}\n${e.stack}`)
+  }
+
 }
 
 //Update a house object, and setting the updated with a new timestamp
 export async function updateHouseById(house){
-  await updateDoc(doc(database, collectionName, house.id), {
-    name: house.name,
-    description: house.description,
-    updated: getReadableTimestamp()
-  });
+  try{
+    await updateDoc(doc(database, collectionName, house.id), {
+      name: house.name,
+      description: house.description,
+      updated: getReadableTimestamp()
+    });
+  }catch (e) {
+    throw new Error(`Failed to update house with id: ${house.id}\n${e.stack}`)
+  }
 }
 
 //Take an houseId and delete housedocument
 export async function deleteHouseById(houseId){
+  try{
   await deleteDoc(doc(database, collectionName, houseId));
+  }catch (e) {
+    throw new Error(`Failed to delete house with id: ${houseId}\n${e.stack}`)
+  }
 }
